Functionality Import products has been done correctly

+++ src/model/SortProductsByPrice.java
@@ -7,7 +7,7 @@ public class SortProductsByPrice implements Comparator<Product> {
 	@Override
 	public int compare( Product o1,  Product o2) {
 		
-		return Integer.parseInt(o1.getPrice()) - Integer.parseInt(o2.getPrice());
+		return (int)Double.parseDouble(o1.getPrice()) - (int)Double.parseDouble(o2.getPrice());
 	}
 
 }
